Split observability status files below size gate
