+++ .pm/tasks/task_43a21163092541809de36036403d7c97.execution.md
@@ -457,3 +457,12 @@ Example:
 - Actual Result: passed; preflight reported source head `1ab48517ccd0c94e9910e520ff80175dedce0ead`, ahead of base `8`, behind base `0`, Pre-PR Local Role Review `passed`, and created PR `https://github.com/eng-cc/oasis7/pull/484`.
 - PR URL: https://github.com/eng-cc/oasis7/pull/484
 - Blocker / Next Action: push this PR evidence entry, then watch PR checks, mergeability, comments, and review threads.
+
+## 2026-06-16 00:10:30 CST / tpm
+- 完成内容: diagnosed and fixed PR `#484` required-gate failure.
+- 遗留事项: GitHub required-gate must rerun on the pushed fix before merge.
+- Action: inspected failing GitHub Actions log for `required-gate`; reproduced the failure signature locally as Rust 1200-line gate overflow; split path observability status structs/helper into `status_payload_observability.rs` and moved the storage challenge degraded readiness regression into a dedicated observability test split.
+- Validation Command: `gh run view 27557394614 --job 81459938073 --log-failed`; `./scripts/check-rust-file-size.sh`; `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_runtime build_chain_status_payload_marks_storage_challenge_network_degraded_not_ready -- --nocapture`; `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_runtime observability -- --nocapture`; `./scripts/cargo-dev.sh fmt`; `git diff --check`
+- Expected Result: file-size gate reports zero oversized Rust files; the moved storage challenge test and full observability filter pass; formatting and whitespace checks stay clean.
+- Actual Result: passed. GitHub failure was `check-rust-file-size.sh` reporting `status_payload.rs` at 1244 lines and `oasis7_chain_runtime_observability_tests.rs` at 1222 lines. After the split, `status_payload.rs` is 1176 lines, `oasis7_chain_runtime_observability_tests.rs` is 1106 lines, and `oasis7_chain_runtime_observability_storage_challenge_tests.rs` is 120 lines. File-size gate passed; exact storage challenge test passed; full observability filter passed 15 tests; fmt and diff check passed.
+- Blocker / Next Action: commit and push the fix, then re-watch PR checks, mergeability, comments, and review threads.
